requirements.txt
pipenv==2022.10.4
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
virtualenv==20.16.5

